Fix sort() expressions
For number compare, 'b > a' is wrong - it should be 'b - a'
- Took away two sort() calls we don't really need (which were wrong too)
- Fixed two compare functions

This makes fewer solutions when using for instance
'node cntdn.js 1 2 3 4 5 6 780'
Before fix, several trivial solutions were added to the list.
The two corrected remaining sort functions makes sure that same-ish
results are sorted regarding unimportant differences in the sequences,
like '6 * 5 * 2' and '5 * 2 * 6' so that they are recognized as same result.