Always give responses on inputs in the 'try' field
is_letters (new) and is_numbers keep track of which puzzle mode we are in.
The 'Try' field code detects which kind of input is entered independent of which
puzzle mode we are in.
If match between the 'Try' field mode and the puzzle mode, then the 'Try' field is
considered to try to solve the puzzle.
If not same mode, the 'Try' is either a formula which is calculated, or a word which is
looked up in the dictionary, or neither if 'Try' has a mix of letters and digits.